Adjusted kafka cluster, added more kafka info & updated makefile
Adjusted statefulset to use the right xtdb-image like in dan's helm-chart. Adjusted the kafka cluster to use the "local:9092". Added a bunch of necessary additional information to the values.yaml which helped to get the xtdb-log to create/connect with the kafka controllers and statefulsets. Adjusted parameters to the gcpbucket to match what's used in the values file, which was causing an error. Created a makefile connect function to easily connect to the xtdb postgre psql wire port and communicate with the cloud supported xtdb instance.