Removed unnecessary sections, working on verification
Deleted TLS stuff that wasn't being used. Deleted a lot of the Makefile with redundant variables, including a lot of parameters that didn't need to be passed in since they were already defined in the values.yaml file. Added a help command to help people unfamiliar with the makefile use it. Added an mpdev verify function in makefile. removed a large portion of the schema since it was all stuff that the user in the marketplace didn't need to be able to change, so it didn't need to be shown...