Adjusted Marketplace MakeFile, created README
I adjusted the makefile in the new GCMarketplace folder to better reflect the experience that we want someone downloading/using XTDB to have on the marketplace. This involves asking the user for their own input for various parts, and adding temp files to store this (that are deleted when they need to be cleared up...). I also created a README file that reflects how postgresql created theirs, it fundamentally goes through the process, but I want to adjust it to specifically help perform the pre-requisites for the click-to-deploy next...