Dockerfile for creating a secure instance of XTDB
This dockerfile will create a docker image that wraps the original xtdb image in a script that generates SSL certificates, converts them into a JKS, and then puts the resultant config file into XTDB, along with needing to create a password. This ensure that this version of XTDB is created as securely as possible! I have updated the makefile to ensure it is easy to generate, run, connect to and cleanup...