Refactored Makefile + README changes
I have entirely changed the Makefile to have some hidden from the consumer Dev functionality, to make testing on the folder easier. I have also done lots of formatting changes to ensure that it is easily readable for anyone else seeing it, and can be followed as such. I deleted an un-needed txt file, and I updated the README to include a section on the Compute Engine storage disks that sometimes weren't getting deleted properly...